intelli-surveillance/pyproject.toml
[project]
name = "intelli-surveillance"
version = "0.1.0"
description = "Système de Surveillance Intelligente Multimodale pour la Prévention du Vol en Grande Distribution"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.10.1",
    "aiofiles>=25.1.0",
    "alembic>=1.18.3",
    "asyncpg>=0.31.0",
    "bitsandbytes>=0.49.1",
    "celery[redis]>=5.3.1",
    "fastapi>=0.128.5",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "numpy>=2.3.2",
    "opencv-python>=4.13.0.92",
    "orjson>=3.11.7",
    "pillow>=11.3.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.2",
    "qwen-vl-utils>=0.0.14",
    "redis>=7.1.0",
    "rich>=14.3.2",
    "safetensors>=0.6.2",
    "scikit-image>=0.26.0",
    "sqlalchemy[asyncio]>=2.0.46",
    "transformers>=4.55.4",
    "ultralytics>=8.4.13",
    "uvicorn[standard]>=0.40.0",
    "websockets>=16.0",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.0",
]
